+++ scout/adapter/mongo/case.py
@@ -501,7 +501,7 @@ class CaseHandler(object):
         result = self.case_collection.delete_one(query)
         return result
 
-    def load_case(self, config_data, update=False):
+    def load_case(self, config_data, update=False, keep_actions=True):
         """Load a case into the database
 
         Check if the owner and the institute exists.
@@ -509,7 +509,7 @@ class CaseHandler(object):
         Args:
             config_data(dict): A dictionary with all the necessary information
             update(bool): If existing case should be updated
-
+            keep_actions(bool): Attempt transfer of existing case user actions to new vars
         Returns:
             case_obj(dict)
         """
@@ -528,6 +528,7 @@ class CaseHandler(object):
         old_caseid = "-".join([case_obj["owner"], case_obj["display_name"]])
         old_case = self.case(old_caseid)
         # This is to keep sanger order and validation status
+
         old_sanger_variants = self.case_sanger_variants(case_obj["_id"])
 
         if old_case:
@@ -543,6 +544,13 @@ class CaseHandler(object):
         existing_case = self.case(case_obj["_id"])
         if existing_case and not update:
             raise IntegrityError("Case %s already exists in database" % case_obj["_id"])
+
+        old_evaluated_variants = (
+            None  # acmg, manual rank, cancer tier, dismissed, mosaic, commented
+        )
+        if existing_case and keep_actions:
+            # collect all variants with user actions for this case
+            old_evaluated_variants = list(self.evaluated_variants(case_obj["_id"]))
 
         files = [
             {"file_name": "vcf_snv", "variant_type": "clinical", "category": "snv"},
@@ -585,7 +593,7 @@ class CaseHandler(object):
         except (IntegrityError, ValueError, ConfigError, KeyError) as error:
             LOG.warning(error)
 
-        if existing_case and update:
+        if existing_case:
             case_obj["rerun_requested"] = False
             if case_obj["status"] in ["active", "archived"]:
                 case_obj["status"] = "inactive"
@@ -596,6 +604,11 @@ class CaseHandler(object):
             self.update_case_sanger_variants(
                 institute_obj, case_obj, old_sanger_variants
             )
+
+            if keep_actions and old_evaluated_variants:
+                self.update_variant_actions(
+                    institute_obj, case_obj, old_evaluated_variants
+                )
 
         else:
             LOG.info("Loading case %s into database", case_obj["display_name"])