+++ CHANGELOG.md
@@ -8,6 +8,9 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 - Clinvar CNVs IGV track
+- Gene panels can have maintainers
+- Keep variant actions (dismissed, manual rank, mosaic, acmg, comments) upon variant re-upload
+- Keep variant actions also on full case re-upload
 
 ### Fixed
 - Fix the link to Ensembl for SV variants when genome build 38.
@@ -18,8 +21,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Updated to a version of IGV where bigBed tracks are visualized correctly
 - Clinvar submission files are named according to the content (variant_data and case_data)
 - Always show causatives from other cases in case overview
+- Correct disease associations for gene symbol aliases that exist as separate genes
 - Re-add "custom annotations" for SV variants
-
 
 ### Changed
 - Runs all CI checks in github actions
@@ -37,8 +40,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Update HPO terms using downloaded definitions files
 - Option to use alternative flask config when running `scout serve`
 - Requirement to use loqusdb >= 2.5 if integrated
-- Gene panels can have maintainers
-- Keep variant actions (dismissed, manual rank, mosaic, acmg, comments) upon variant re-upload
 
 ### Fixed
 - Do not display Pedigree panel in cancer view